seperate the tests based on the file seperation
