+++ shared-data/js/__tests__/fixtures.test.ts
@@ -2,8 +2,6 @@ import { useTranslation } from 'react-i18next'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 import {
-  ABSORBANCE_READER_V1_FIXTURE,
-  FAKE_STAGING_AREA_RIGHT_SLOT,
   FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE,
   FAKE_WASTE_CHUTE_WITH_EMPTY_SLOT_FIXTURE,
   FLEX_STACKER_V1_FIXTURE,
@@ -12,9 +10,6 @@ import {
   FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_NO_COVER_FIXTURE,
   HEATERSHAKER_MODULE_V1_FIXTURE,
   MAGNETIC_BLOCK_V1_FIXTURE,
-  SINGLE_CENTER_SLOT_FIXTURE,
-  SINGLE_LEFT_SLOT_FIXTURE,
-  SINGLE_RIGHT_SLOT_FIXTURE,
   STAGING_AREA_RIGHT_SLOT_FIXTURE,
   STAGING_AREA_SLOT_WITH_MAGNETIC_BLOCK_V1_FIXTURE,
   STAGING_AREA_SLOT_WITH_WASTE_CHUTE_RIGHT_ADAPTER_COVERED_FIXTURE,
@@ -42,13 +37,6 @@ import {
   replaceCutoutFixtureWithComboFixture,
 } from '../fixtures'
 import { getDeckDefFromRobotType } from '../helpers'
-import {
-  getCutoutFixtureReplacementIfNeeded,
-  getReplacementFixtureForFakeFixture,
-  getReplacementFixtureForFixtureRemoval,
-  replaceCutoutFixtureForFixtureRemoval,
-} from '../helpers/deckConfiguration/getFixtureFrom'
-import { getVisualSlotIdFromAAId } from '../helpers/deckConfiguration/getVisualSlotFrom'
 
 import type { Mock } from 'vitest'
 import type { CutoutConfig, DeckConfiguration } from '..'
@@ -209,90 +197,6 @@ describe('getAddressableAreaWithFakesMatchForAreaId', () => {
   })
 })
 
-describe('getCutoutFixtureReplacementIfNeeded', () => {
-  it('Should get FAKE_STAGING_AREA_RIGHT_SLOT if matches condition', () => {
-    const result = getCutoutFixtureReplacementIfNeeded(
-      SINGLE_RIGHT_SLOT_FIXTURE,
-      'cutoutA3',
-      deckDef
-    )
-
-    expect(result).toEqual(FAKE_STAGING_AREA_RIGHT_SLOT)
-  })
-
-  it('Should return current cutoutfixtureid when does not match condition', () => {
-    const result = getCutoutFixtureReplacementIfNeeded(
-      STAGING_AREA_RIGHT_SLOT_FIXTURE,
-      'cutoutB3',
-      deckDef
-    )
-
-    expect(result).toEqual(STAGING_AREA_RIGHT_SLOT_FIXTURE)
-  })
-})
-
-describe('getReplacementFixtureForFixtureRemoval', () => {
-  it('Should get SINGLE_RIGHT_SLOT_FIXTURE to replace staging slot removal', () => {
-    const result = getReplacementFixtureForFixtureRemoval(
-      STAGING_AREA_RIGHT_SLOT_FIXTURE,
-      'cutoutD3',
-      'fakeD4'
-    )
-
-    expect(result).toEqual(SINGLE_RIGHT_SLOT_FIXTURE)
-  })
-
-  it('Should return SINGLE_RIGHT_SLOT_FIXTURE when using a single right cutout', () => {
-    const result = getReplacementFixtureForFixtureRemoval(
-      SINGLE_RIGHT_SLOT_FIXTURE,
-      'cutoutD3',
-      'D3'
-    )
-
-    expect(result).toEqual(SINGLE_RIGHT_SLOT_FIXTURE)
-  })
-
-  it('Should return SINGLE_LEFT_SLOT_FIXTURE when using a single right cutout', () => {
-    const result = getReplacementFixtureForFixtureRemoval(
-      SINGLE_RIGHT_SLOT_FIXTURE,
-      'cutoutA1',
-      'A1'
-    )
-
-    expect(result).toEqual(SINGLE_LEFT_SLOT_FIXTURE)
-  })
-
-  it('Should return SINGLE_CENTER_SLOT_FIXTURE when using a single center cutout', () => {
-    const result = getReplacementFixtureForFixtureRemoval(
-      SINGLE_RIGHT_SLOT_FIXTURE,
-      'cutoutA2',
-      'A2'
-    )
-
-    expect(result).toEqual(SINGLE_CENTER_SLOT_FIXTURE)
-  })
-
-  it('Should return staging area fixture instead of combo', () => {
-    const result = getReplacementFixtureForFixtureRemoval(
-      STAGING_AREA_SLOT_WITH_MAGNETIC_BLOCK_V1_FIXTURE,
-      'cutoutD3',
-      'magneticBlockV1D3'
-    )
-
-    expect(result).toEqual(STAGING_AREA_RIGHT_SLOT_FIXTURE)
-  })
-
-  it('Should return stacker instead of stacker mag block combo', () => {
-    const result = getReplacementFixtureForFixtureRemoval(
-      FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
-      'cutoutD3',
-      'magneticBlockV1D3'
-    )
-
-    expect(result).toEqual(FLEX_STACKER_V1_FIXTURE)
-  })
-})
-
 describe('getFlexDeckDefAAByFixtureIdForCutoutId', () => {
   it('Should return a dic of fixtures and aa for cutoutA3', () => {
     const cutoutA3Result = getAAsToFixtureIdFromDeckDefWithFakes(
@@ -399,29 +303,6 @@ describe('getFlexDeckDefAAByFixtureIdForCutoutId', () => {
   })
 })
 
-describe('getReplacementFixtureForFakeFixture', () => {
-  it('should return staging slot fixture instead of FAKE_STAGING_AREA_RIGHT_SLOT', () => {
-    const result = getReplacementFixtureForFakeFixture(
-      FAKE_STAGING_AREA_RIGHT_SLOT
-    )
-    expect(result).toEqual(SINGLE_RIGHT_SLOT_FIXTURE)
-  })
-
-  it('should return waste chute fixture instead of FAKE_WASTE_CHUTE_WITH_EMPTY_SLOT', () => {
-    const result = getReplacementFixtureForFakeFixture(
-      FAKE_WASTE_CHUTE_WITH_EMPTY_SLOT_FIXTURE
-    )
-    expect(result).toEqual(WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE)
-  })
-
-  it('should return mag block fixture instead of FAKE_STAGING_SLOT_WITH_MAG_BLOCK', () => {
-    const result = getReplacementFixtureForFakeFixture(
-      FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE
-    )
-    expect(result).toEqual(MAGNETIC_BLOCK_V1_FIXTURE)
-  })
-})
-
 describe('getVisualSlotIdForAA', () => {
   it('should get vs name for single right slot', () => {
     const result = getVisualSlotIdForAA(
@@ -448,61 +329,6 @@ describe('getVisualSlotIdForAA', () => {
       'magneticBlockV1D2'
     )
     expect(result).toEqual('VSD2')
-  })
-})
-
-describe('replaceCutoutFixtureRemove', () => {
-  it('should get fixture replacment for FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE', () => {
-    const result = replaceCutoutFixtureForFixtureRemoval(
-      FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
-      'cutoutD3',
-      'magneticBlockV1D3'
-    )
-    expect(result).toEqual(FLEX_STACKER_V1_FIXTURE)
-  })
-
-  it('should get fixture replacment for FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE without a stacker', () => {
-    const result = replaceCutoutFixtureForFixtureRemoval(
-      FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
-      'cutoutC3',
-      'flexStackerModuleV1C4'
-    )
-    expect(result).toEqual(FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE)
-  })
-
-  it('should get fixture replacment for MAGNETIC_BLOCK_V1_FIXTURE', () => {
-    const result = replaceCutoutFixtureForFixtureRemoval(
-      MAGNETIC_BLOCK_V1_FIXTURE,
-      'cutoutD3',
-      'magneticBlockV1D3'
-    )
-    expect(result).toEqual(SINGLE_RIGHT_SLOT_FIXTURE)
-  })
-  it('should get fixture replacment for FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_COVERED_FIXTURE', () => {
-    const result = replaceCutoutFixtureForFixtureRemoval(
-      FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_NO_COVER_FIXTURE,
-      'cutoutD3',
-      'flexStackerModuleV1D4'
-    )
-    expect(result).toEqual(WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE)
-  })
-
-  it('should get fixture replacment for FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_COVERED_FIXTURE when waste chute is on removed', () => {
-    const result = replaceCutoutFixtureForFixtureRemoval(
-      FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_NO_COVER_FIXTURE,
-      'cutoutD3',
-      '1ChannelWasteChute'
-    )
-    expect(result).toEqual(FLEX_STACKER_V1_FIXTURE)
-  })
-
-  it('should get fixture replacment for ABSORBANCE_READER_V1_FIXTURE', () => {
-    const result = replaceCutoutFixtureForFixtureRemoval(
-      ABSORBANCE_READER_V1_FIXTURE,
-      'cutoutD3',
-      'absorbanceReaderV1D3'
-    )
-    expect(result).toEqual(SINGLE_RIGHT_SLOT_FIXTURE)
   })
 })
 
@@ -589,18 +415,6 @@ describe('isFixtureInModules', () => {
   it('should return false for mag block fixture', () => {
     const result = isFixtureInUsbModules(MAGNETIC_BLOCK_V1_FIXTURE)
     expect(result).toEqual(false)
-  })
-})
-
-describe('getVisualSlotIdFromAAId', () => {
-  it('should get VSD4 for flexStackerModuleV1D4', () => {
-    const vs = getVisualSlotIdFromAAId('flexStackerModuleV1D4')
-    expect(vs).toEqual('VSD4')
-  })
-
-  it('should get VSD3 for D3', () => {
-    const vs = getVisualSlotIdFromAAId('D3')
-    expect(vs).toEqual('VSD3')
   })
 })
